inline the sdk provider
