--- enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk/components/SdkIframeEmbedProvider/index.ts
@@ -1,1 +1,0 @@
-export { SdkIframeEmbedProvider } from "./SdkIframeEmbedProvider";
