+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk/components/SdkIframeEmbedRoute/SdkIframeEmbedRoute.module.css
@@ -1,4 +1,4 @@
-.InteractiveEmbeddingPortalContainer {
+.SdkIframeEmbedPortalContainer {
   position: absolute;
   z-index: var(--mb-overlay-z-index);
 }
