+++ frontend/src/metabase/embedding-sdk/theme/default-component-theme.ts
@@ -64,6 +64,10 @@ export const DEFAULT_METABASE_COMPONENT_THEME: MetabaseComponentTheme = {
   },
   question: {
     backgroundColor: "transparent",
+
+    toolbar: {
+      backgroundColor: "var(--mb-color-background)",
+    },
   },
 
   table: {
